requirements.txt
langchain>=0.0.262
langchain-community>=0.1.3
pydantic>=2.0
python-dotenv>=1.0.0
azure-ai-openai>=1.0.0
langchain_sql_database>=0.0.1
sqlalchemy>=1.4.0
langchain-openai>=0.3.0
